# apps/api/openapi.json
{
  "components": {
    "schemas": {
      "AnalyticsRunRequest": {
        "additionalProperties": false,
        "properties": {
          "stages": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stages"
          }
        },
        "title": "AnalyticsRunRequest",
        "type": "object"
      },
      "AnalyticsRunResult": {
        "additionalProperties": false,
        "properties": {
          "job_id": {
            "title": "Job Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "job_id",
          "status"
        ],
        "title": "AnalyticsRunResult",
        "type": "object"
      },
...
      "AssistantDraftRequest": {
        "additionalProperties": false,
        "properties": {
          "facts": {
            "items": {
              "$ref": "#/components/schemas/ReportFact"
            },
            "title": "Facts",
            "type": "array"
          }
        },
        "required": [
          "facts"
        ],
        "title": "AssistantDraftRequest",
        "type": "object"
      },
      "AssistantQueryRequest": {
        "additionalProperties": false,
        "properties": {
          "question": {
            "title": "Question",
            "type": "string"
          }
        },
        "required": [
          "question"
        ],
        "title": "AssistantQueryRequest",
        "type": "object"
      },
      "AssistantQueryResponse": {
        "additionalProperties": false,
        "properties": {
          "filter": {
            "$ref": "#/components/schemas/EvidenceSearchFilter"
          },
          "result_count": {
            "title": "Result Count",
            "type": "integer"
          },
          "results": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Results",
            "type": "array"
          }
        },
        "required": [
          "filter",
          "result_count",
          "results"
        ],
        "title": "AssistantQueryResponse",
        "type": "object"
      },
      "AuditEntry": {
        "additionalProperties": false,
        "properties": {
          "action": {
            "title": "Action",
            "type": "string"
          },
          "actor": {
            "title": "Actor",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "entry_hash": {
            "title": "Entry Hash",
            "type": "string"
          },
          "object_id": {
            "title": "Object Id",
            "type": "string"
          },
          "object_type": {
            "title": "Object Type",
            "type": "string"
          },
          "payload_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payload Sha256"
          },
          "prev_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prev Hash"
          },
          "role": {
            "title": "Role",
            "type": "string"
          },
          "seq": {
            "title": "Seq",
            "type": "integer"
          },
          "signature": {
            "title": "Signature",
            "type": "string"
          },
          "ts_utc": {
            "title": "Ts Utc",
            "type": "string"
          }
        },
        "required": [
          "seq",
          "prev_hash",
          "entry_hash",
          "ts_utc",
          "actor",
          "role",
          "action",
          "object_type",
          "object_id",
          "payload_sha256",
          "details",
          "signature"
        ],
        "title": "AuditEntry",
        "type": "object"
      },
      "AuditVerifyResult": {
        "additionalProperties": false,
        "properties": {
          "first_bad_seq": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "First Bad Seq"
          },
          "head_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Head Hash"
          },
          "length": {
            "title": "Length",
            "type": "integer"
          },
          "ok": {
            "title": "Ok",
            "type": "boolean"
          }
        },
        "required": [
          "ok",
          "length",
          "head_hash",
          "first_bad_seq"
        ],
        "title": "AuditVerifyResult",
        "type": "object"
      },
...
      "Detection": {
        "additionalProperties": false,
        "properties": {
          "bbox": {
            "maxItems": 4,
            "minItems": 4,
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "title": "Bbox",
            "type": "array"
          },
          "cls": {
            "enum": [
              "person",
              "vehicle",
              "two_wheeler",
              "other"
            ],
            "title": "Cls",
            "type": "string"
          },
          "derived_from": {
            "title": "Derived From",
            "type": "string"
          },
          "frame_id": {
            "title": "Frame Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "score": {
            "title": "Score",
            "type": "number"
          }
        },
        "required": [
          "id",
          "frame_id",
          "cls",
          "score",
          "bbox",
          "model",
          "derived_from"
        ],
        "title": "Detection",
        "type": "object"
      },
      "DraftSentence": {
        "additionalProperties": false,
        "properties": {
          "evidence_ids": {
            "items": {
              "type": "string"
            },
            "title": "Evidence Ids",
            "type": "array"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "text",
          "evidence_ids"
        ],
        "title": "DraftSentence",
        "type": "object"
      },
...
      "EvidenceSearchFilter": {
        "additionalProperties": false,
        "properties": {
          "channels": {
            "anyOf": [
              {
                "items": {
                  "type": "integer"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Channels"
          },
          "deleted_only": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Deleted Only"
          },
          "detection_class": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detection Class"
          },
          "from_ist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "From Ist"
          },
          "log_kind": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Log Kind"
          },
          "motion_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Motion Min"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "to_ist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "To Ist"
          }
        },
        "title": "EvidenceSearchFilter",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "title": "Password",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "Me": {
        "additionalProperties": false,
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "role": {
            "enum": [
              "examiner",
              "reviewer",
              "admin"
            ],
            "title": "Role",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "username",
          "role",
          "display_name"
        ],
        "title": "Me",
        "type": "object"
      },
      "MotionSegment": {
        "additionalProperties": false,
        "properties": {
          "channel": {
            "title": "Channel",
            "type": "integer"
          },
          "end_norm_us": {
            "title": "End Norm Us",
            "type": "integer"
          },
          "frames": {
            "title": "Frames",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "image_id": {
            "title": "Image Id",
            "type": "string"
          },
          "peak_score": {
            "title": "Peak Score",
            "type": "number"
          },
          "start_norm_us": {
            "title": "Start Norm Us",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "image_id",
          "channel",
          "start_norm_us",
          "end_norm_us",
          "peak_score",
          "frames"
        ],
        "title": "MotionSegment",
        "type": "object"
      },
      "Page_AuditEntry_": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AuditEntry"
            },
            "title": "Items",
            "type": "array"
          },
          "next_cursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Next Cursor"
          }
        },
        "required": [
          "items",
          "next_cursor"
        ],
        "title": "Page[AuditEntry]",
        "type": "object"
      },
...
      "ReportFact": {
        "additionalProperties": false,
        "properties": {
          "id": {
            "title": "Id",
            "type": "string"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "text"
        ],
        "title": "ReportFact",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Multi-vendor DVR/NVR forensic platform \u2014 API (docs/02-BACKEND.md).",
    "title": "Pramaan API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/auth/login": {
      "post": {
        "operationId": "login_api_auth_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Me"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Login",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/cases/{cid}/analytics/run": {
      "post": {
        "operationId": "run_analytics_api_cases__cid__analytics_run_post",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AnalyticsRunRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsRunResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Analytics",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/cases/{cid}/assistant/draft": {
      "post": {
        "operationId": "assistant_draft_api_cases__cid__assistant_draft_post",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantDraftRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DraftSentence"
                  },
                  "title": "Response Assistant Draft Api Cases  Cid  Assistant Draft Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assistant Draft",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/cases/{cid}/assistant/query": {
      "post": {
        "operationId": "assistant_query_api_cases__cid__assistant_query_post",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssistantQueryRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AssistantQueryResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Assistant Query",
        "tags": [
          "assistant"
        ]
      }
    },
    "/api/cases/{cid}/audit": {
      "get": {
        "operationId": "list_audit_api_cases__cid__audit_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_AuditEntry_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Audit",
        "tags": [
          "audit"
        ]
      }
    },
    "/api/cases/{cid}/audit/verify": {
      "get": {
        "operationId": "verify_audit_api_cases__cid__audit_verify_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuditVerifyResult"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Verify Audit",
        "tags": [
          "audit"
        ]
      }
    },
...
    "/api/cases/{cid}/detections": {
      "get": {
        "operationId": "list_detections_api_cases__cid__detections_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/Detection"
                  },
                  "title": "Response List Detections Api Cases  Cid  Detections Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Detections",
        "tags": [
          "analytics"
        ]
      }
    },
...
    "/api/cases/{cid}/motion": {
      "get": {
        "operationId": "list_motion_api_cases__cid__motion_get",
        "parameters": [
          {
            "in": "path",
            "name": "cid",
            "required": true,
            "schema": {
              "title": "Cid",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "channel",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Channel"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MotionSegment"
                  },
                  "title": "Response List Motion Api Cases  Cid  Motion Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Motion",
        "tags": [
          "analytics"
        ]
      }
    },
...
  }
}